+++ builds/4chan-X.user.js
@@ -18,7 +18,7 @@
 // @icon         data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAADAAAAAwAgMAAAAqbBEUAAAACVBMVEUAAGcAAABmzDNZt9VtAAAAAXRSTlMAQObYZgAAAHFJREFUKFOt0LENACEIBdBv4Qju4wgWanEj3D6OcIVMKaitYHEU/jwTCQj8W75kiVCSBvdQ5/AvfVHBin11BgdRq3ysBgfwBDRrj3MCIA+oAQaku/Q1cNctrAmyDl577tOThYt/Y1RBM4DgOHzM0HFTAyLukH/cmRnqAAAAAElFTkSuQmCC
 // ==/UserScript==
 /*
-* 4chan X - Version 1.1.2 - 2013-04-27
+* 4chan X - Version 1.1.2 - 2013-04-28
 *
 * Licensed under the MIT license.
 * https://github.com/seaweedchan/4chan-x/blob/master/LICENSE
@@ -6689,6 +6689,8 @@
         if (input.name === 'Scroll BG') {
           $.on(input, 'change', ThreadUpdater.cb.scrollBG);
           ThreadUpdater.cb.scrollBG();
+        } else if (input.name === 'Auto Update') {
+          $.on(input, 'change', ThreadUpdater.update);
         }
         subEntries.push({
           el: el
@@ -6784,7 +6786,8 @@
         if (ThreadUpdater.online) {
           return ThreadUpdater.timeoutID = setTimeout(ThreadUpdater.timeout, 1000);
         } else {
-          return clearTimeout(ThreadUpdater.timeoutID);
+          clearTimeout(ThreadUpdater.timeoutID);
+          return ThreadUpdater.set('timer', 'Update');
         }
       },
       interval: function() {
@@ -6806,7 +6809,9 @@
             g.DEAD = false;
             ThreadUpdater.parse(JSON.parse(req.response).posts);
             ThreadUpdater.lastModified = req.getResponseHeader('Last-Modified');
-            ThreadUpdater.set('timer', ThreadUpdater.getInterval());
+            if (Conf['Auto Update']) {
+              ThreadUpdater.set('timer', ThreadUpdater.getInterval());
+            }
             break;
           case 404:
             g.DEAD = true;
@@ -6820,8 +6825,10 @@
             });
             break;
           default:
-            ThreadUpdater.outdateCount++;
-            ThreadUpdater.set('timer', ThreadUpdater.getInterval());
+            if (Conf['Auto Update']) {
+              ThreadUpdater.outdateCount++;
+              ThreadUpdater.set('timer', ThreadUpdater.getInterval());
+            }
             /*
             Status Code 304: Not modified
             By sending the `If-Modified-Since` header we get a proper status code, and no response.
@@ -6887,7 +6894,11 @@
         return;
       }
       ThreadUpdater.seconds = 0;
-      ThreadUpdater.set('timer', '...');
+      if (Conf['Auto Update']) {
+        ThreadUpdater.set('timer', '...');
+      } else {
+        ThreadUpdater.set('timer', 'Update');
+      }
       if (ThreadUpdater.req) {
         ThreadUpdater.req.onloadend = null;
         ThreadUpdater.req.abort();